x/epoch/genesis_test.go
package epoch_test

import (
	"testing"

	keepertest "github.com/platine-network/platine/testutil/keeper"
	"github.com/platine-network/platine/testutil/nullify"
	"github.com/platine-network/platine/x/epoch"
	"github.com/platine-network/platine/x/epoch/types"
	"github.com/stretchr/testify/require"
)

func TestGenesis(t *testing.T) {
	genesisState := types.GenesisState{
		Params: types.DefaultParams(),

		EpochList: []types.Epoch{
			{
				Id: 0,
			},
			{
				Id: 1,
			},
		},
		EpochCount: 2,
		// this line is used by starport scaffolding # genesis/test/state
	}

	k, ctx := keepertest.EpochKeeper(t)
	epoch.InitGenesis(ctx, *k, genesisState)
	got := epoch.ExportGenesis(ctx, *k)
	require.NotNil(t, got)

	nullify.Fill(&genesisState)
	nullify.Fill(got)

	require.ElementsMatch(t, genesisState.EpochList, got.EpochList)
	require.Equal(t, genesisState.EpochCount, got.EpochCount)
	// this line is used by starport scaffolding # genesis/test/assert
}
